+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Make `DebugImagesLoader` public ([#2993](https://github.com/getsentry/sentry-java/pull/2993))
 
+### Fixes
+
+- Set export flag on Android receivers on API 33+ ([#2990](https://github.com/getsentry/sentry-java/pull/2990))
+
 ## 6.31.0
 
 ### Features